% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSentences.R
\name{compareSentences}
\alias{compareSentences}
\title{Flag differences between two text excerpts}
\usage{
compareSentences(
  sentence1,
  sentence2,
  highlight_diff = "underline",
  NA.string = not.applicable
)
}
\arguments{
\item{sentence1}{Original text}

\item{sentence2}{Revised text}

\item{highlight_diff}{Font type used to highlight text differences. Defaults to \code{underline}. Other options are \code{bold}, \code{italics}, and \code{standard}.}

\item{NA.string}{String defining characters recognised as "NA".}
}
\value{
A LaTeX formatted string showing text differences when rendered.
}
\description{
This function produces a LaTeX mark up that highlights text differences between two input sentences.
}
\examples{

library(respondR)
library(tidyverse)

compareSentences(sentence1 = "The skye is blue and the grass is green.", 
                sentence2 = "The ocean is blue and the apples are green!",
                highlight_diff = "underline", 
                NA.string = c("n/a", "na", "N/A", "NA", "Not applicable",
                 "not applicable", "Not Applicable"))

compareSentences(sentence1 = "My name is Patricia.", 
                sentence2 = "My name is Stacey.", 
                highlight_diff = "underline", 
                NA.string = c("n/a", "na", "N/A", "NA"))

compareSentences(sentence1 = "My tailor is rich, 
                and those flowers are beautiful!", 
                sentence2 = "Their tailor is very wealthy, 
                and these flowers are beautiful!",
                highlight_diff = "underline", 
                NA.string = c("n/a", "na", "N/A", "NA"))
}
\author{
Phil J. Bouchet
}
