% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_doc.R
\name{create_doc}
\alias{create_doc}
\title{Copy the Microsoft Word template document}
\usage{
create_doc(rstudio = TRUE, destination = NULL, file.name = NULL)
}
\arguments{
\item{rstudio}{Logical. Whether the function is being called from within RStudio. Defaults to TRUE.}

\item{destination}{Destination folder. If \code{rstudio = TRUE}, the file will be saved in the same directory as the current file. If \code{rstudio = TRUE} and \code{destination} is not specified, will default to the current working directory.}
}
\value{
A LaTeX formatted string showing text differences when rendered.
}
\description{
Copy the Microsoft Word template document
}
\author{
Phil J. Bouchet
}
