% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_template.R
\name{initialise_template}
\alias{initialise_template}
\title{Download Microsoft Word template document}
\usage{
initialise_template(rstudio = TRUE, destination = NULL)
}
\arguments{
\item{rstudio}{Logical. Whether the function is being called from within RStudio Defaults to TRUE.}

\item{destination}{Destination folder. If \code{rstudio = TRUE}, the file will be saved in the same directory as the current file. Otherwise, defaults to the current working directory.}
}
\value{
A LaTeX formatted string showing text differences when rendered.
}
\description{
Download Microsoft Word template document
}
\author{
Phil J. Bouchet
}
