% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{response}
\alias{response}
\title{Output formats for respondR template}
\usage{
response(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents in a suitable style
}
\author{
Phil J. Bouchet
}
