% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqle.R
\name{seqle}
\alias{seqle}
\title{Look for sequences of integers in a vector}
\usage{
seqle(x)
}
\arguments{
\item{x}{Vector of integers}
}
\value{
A list with two components:
\tabular{ll}{
\code{lengths} \tab An integer vector containing the length of each run.\cr
\code{values} \tab a vector of the same length as lengths with the corresponding values.\cr
}
}
\description{
The function is adapted from the \code{rle} function for computing the lengths and values of runs of equal values in a vector -- or the reverse operation.
}
\references{
https://stackoverflow.com/questions/8400901/group-integer-vector-into-consecutive-runs/8402950#8402950
}
\author{
Carl Witthoft
}
